작업환경 변경

+++ trade/src/main/java/com/seomse/stock/trade/HoldStock.java
@@ -72,7 +72,7 @@ public class HoldStock implements Stock {
      */
     public double getEvaluationAmount(double price ,double fee){
 
-        double total  = price * count;
+        double total  = price * (double)count;
         total -= total*fee;
         return total;
     }
